+++ db-service/lib/cqn2sql.js
@@ -8,7 +8,7 @@ const { Readable } = require('stream')
 
 const DEBUG = cds.debug('sql|sqlite')
 const LOG_SQL = cds.log('sql')
-const LOG_SQLITE =  cds.log('sqlite')
+const LOG_SQLITE = cds.log('sqlite')
 
 class CQN2SQLRenderer {
   /**
@@ -21,10 +21,12 @@ class CQN2SQLRenderer {
     this.class = new.target // for IntelliSense
     this.class._init() // is a noop for subsequent calls
     this.model = srv?.model
-
     // Overwrite smart quoting
     if (cds.env.sql.names === 'quoted') {
-      this.class.prototype.name = (name) => name.id || name
+      this.class.prototype.name = (name, query) => {
+        const e = name.id || name
+        return (query?.target || this.model?.definitions[e])?.['@cds.persistence.name'] || e
+      }
       this.class.prototype.quote = (s) => `"${String(s).replace(/"/g, '""')}"`
     }
   }
@@ -84,8 +86,8 @@ class CQN2SQLRenderer {
     if (vars && Object.keys(vars).length && !this.values?.length) this.values = vars
     const sanitize_values = process.env.NODE_ENV === 'production' && cds.env.log.sanitize_values !== false
 
-    
-    if (LOG_SQL._debug || LOG_SQLITE._debug) {
+
+    if (DEBUG && (LOG_SQL._debug || LOG_SQLITE._debug)) {
       let values = sanitize_values && (this.entries || this.values?.length > 0) ? ['***'] : this.entries || this.values || []
       if (values && !Array.isArray(values)) {
         values = [values]
@@ -93,7 +95,7 @@ class CQN2SQLRenderer {
       DEBUG(this.sql, ...values)
     }
 
-    
+
     return this
   }
 
@@ -124,7 +126,7 @@ class CQN2SQLRenderer {
       target = typeof entity === 'string' ? { name: entity } : q.CREATE.entity
     }
 
-    const name = this.name(target.name)
+    const name = this.name(target.name, q)
     // Don't allow place holders inside views
     delete this.values
     this.sql =
@@ -195,6 +197,7 @@ class CQN2SQLRenderer {
     Association: () => false,
     Composition: () => false,
     array: () => 'NCLOB',
+    Map: () => 'NCLOB',
     // HANA types
     'cds.hana.TINYINT': () => 'TINYINT',
     'cds.hana.REAL': () => 'REAL',
@@ -213,7 +216,7 @@ class CQN2SQLRenderer {
     const { target } = q
     const isView = target?.query || target?.projection || q.DROP.view
     const name = target?.name || q.DROP.table?.ref?.[0] || q.DROP.view?.ref?.[0]
-    return (this.sql = `DROP ${isView ? 'VIEW' : 'TABLE'} IF EXISTS ${this.quote(this.name(name))}`)
+    return (this.sql = `DROP ${isView ? 'VIEW' : 'TABLE'} IF EXISTS ${this.quote(this.name(name, q))}`)
   }
 
   // SELECT Statements ------------------------------------------------
@@ -352,15 +355,10 @@ class CQN2SQLRenderer {
     const _aliased = as ? s => s + ` as ${this.quote(as)}` : s => s
     if (ref) {
       let z = ref[0]
-      if (cds.env.sql.names === 'quoted') {
-        // use SELECT.from to infer query, cds.infer also expects a query
-        const { target } = q || SELECT.from(from)
-        z = target?.['@cds.persistence.name'] || ref[0]
-      }
       if (z.args) {
-        return _aliased(`${this.quote(this.name(z))}${this.from_args(z.args)}`)
+        return _aliased(`${this.quote(this.name(z, q))}${this.from_args(z.args)}`)
       }
-      return _aliased(this.quote(this.name(z)))
+      return _aliased(this.quote(this.name(z, q)))
     }
     if (from.SELECT) return _aliased(`(${this.SELECT(from)})`)
     if (from.join) return `${this.from(from.args[0])} ${from.join} JOIN ${this.from(from.args[1])}${from.on ? ` ON ${this.where(from.on)}` : ''}`
@@ -423,8 +421,8 @@ class CQN2SQLRenderer {
         ? c =>
           this.expr(c) +
           (c.element?.[this.class._localized] ? ' COLLATE NOCASE' : '') +
-          (c.sort === 'desc' || c.sort === -1 ? ' DESC' : ' ASC')
-        : c => this.expr(c) + (c.sort === 'desc' || c.sort === -1 ? ' DESC' : ' ASC'),
+          (c.sort?.toLowerCase() === 'desc' || c.sort === -1 ? ' DESC' : ' ASC')
+        : c => this.expr(c) + (c.sort?.toLowerCase() === 'desc' || c.sort === -1 ? ' DESC' : ' ASC'),
     )
   }
 
@@ -504,7 +502,7 @@ class CQN2SQLRenderer {
     this.columns = columns
 
     const alias = INSERT.into.as
-    const entity = this.name(q.target?.name || INSERT.into.ref[0])
+    const entity = this.name(q.target?.name || INSERT.into.ref[0], q)
     if (!elements) {
       this.entries = INSERT.entries.map(e => columns.map(c => e[c]))
       const param = this.param.bind(this, { ref: ['?'] })
@@ -531,9 +529,6 @@ class CQN2SQLRenderer {
 
   async *INSERT_entries_stream(entries, binaryEncoding = 'base64') {
     const elements = this.cqn.target?.elements || {}
-    const transformBase64 = binaryEncoding === 'base64'
-      ? a => a
-      : a => a != null ? Buffer.from(a, 'base64').toString(binaryEncoding) : a
     const bufferLimit = 65536 // 1 << 16
     let buffer = '['
 
@@ -564,8 +559,8 @@ class CQN2SQLRenderer {
 
           buffer += '"'
         } else {
-          if (elements[key]?.type in this.BINARY_TYPES) {
-            val = transformBase64(val)
+          if (val != null && elements[key]?.type in this.BINARY_TYPES) {
+            val = Buffer.from(val, 'base64').toString(binaryEncoding)
           }
           buffer += `${keyJSON}${JSON.stringify(val)}`
         }
@@ -583,9 +578,6 @@ class CQN2SQLRenderer {
 
   async *INSERT_rows_stream(entries, binaryEncoding = 'base64') {
     const elements = this.cqn.target?.elements || {}
-    const transformBase64 = binaryEncoding === 'base64'
-      ? a => a
-      : a => a != null ? Buffer.from(a, 'base64').toString(binaryEncoding) : a
     const bufferLimit = 65536 // 1 << 16
     let buffer = '['
 
@@ -612,8 +604,8 @@ class CQN2SQLRenderer {
 
           buffer += '"'
         } else {
-          if (elements[this.columns[key]]?.type in this.BINARY_TYPES) {
-            val = transformBase64(val)
+          if (val != null && elements[this.columns[key]]?.type in this.BINARY_TYPES) {
+            val = Buffer.from(val, 'base64').toString(binaryEncoding)
           }
           buffer += `${sepsub}${val === undefined ? 'null' : JSON.stringify(val)}`
         }
@@ -638,7 +630,7 @@ class CQN2SQLRenderer {
    */
   INSERT_rows(q) {
     const { INSERT } = q
-    const entity = this.name(q.target?.name || INSERT.into.ref[0])
+    const entity = this.name(q.target?.name || INSERT.into.ref[0], q)
     const alias = INSERT.into.as
     const elements = q.elements || q.target?.elements
     const columns = this.columns = INSERT.columns || cds.error`Cannot insert rows without columns or elements`
@@ -684,7 +676,7 @@ class CQN2SQLRenderer {
    */
   INSERT_select(q) {
     const { INSERT } = q
-    const entity = this.name(q.target.name)
+    const entity = this.name(q.target.name, q)
     const alias = INSERT.into.as
     const elements = q.elements || q.target?.elements || {}
     const columns = (this.columns = (INSERT.columns || ObjectKeys(elements)).filter(
@@ -752,8 +744,11 @@ class CQN2SQLRenderer {
       .filter(c => keys.includes(c.name))
       .map(c => `${c.onInsert || c.sql} as ${this.quote(c.name)}`)
 
-    const entity = this.name(q.target?.name || UPSERT.into.ref[0])
-    sql = `SELECT ${managed.map(c => c.upsert)} FROM (SELECT value, ${extractkeys} from json_each(?)) as NEW LEFT JOIN ${this.quote(entity)} AS OLD ON ${keyCompare}`
+    const entity = this.name(q.target?.name || UPSERT.into.ref[0], q)
+    sql = `SELECT ${managed.map(c => c.upsert
+      .replace(/value->/g, '"$$$$value$$$$"->')
+      .replace(/json_type\(value,/g, 'json_type("$$$$value$$$$",'))
+      } FROM (SELECT value as "$$value$$", ${extractkeys} from json_each(?)) as NEW LEFT JOIN ${this.quote(entity)} AS OLD ON ${keyCompare}`
 
     const updateColumns = columns.filter(c => {
       if (keys.includes(c)) return false //> keys go into ON CONFLICT clause
@@ -779,7 +774,7 @@ class CQN2SQLRenderer {
   UPDATE(q) {
     const { entity, with: _with, data, where } = q.UPDATE
     const elements = q.target?.elements
-    let sql = `UPDATE ${this.quote(this.name(entity.ref?.[0] || entity))}`
+    let sql = `UPDATE ${this.quote(this.name(entity.ref?.[0] || entity, q))}`
     if (entity.as) sql += ` AS ${this.quote(entity.as)}`
 
     let columns = []
@@ -811,8 +806,9 @@ class CQN2SQLRenderer {
    * @param {import('./infer/cqn').DELETE} param0
    * @returns {string} SQL
    */
-  DELETE({ DELETE: { from, where } }) {
-    let sql = `DELETE FROM ${this.from(from)}`
+  DELETE(q) {
+    const { DELETE: { from, where } } = q
+    let sql = `DELETE FROM ${this.from(from, q)}`
     if (where) sql += ` WHERE ${this.where(where)}`
     return (this.sql = sql)
   }
@@ -1028,6 +1024,7 @@ class CQN2SQLRenderer {
   /**
    * Calculates the Database name of the given name
    * @param {string|import('./infer/cqn').ref} name
+   * @param {import('./infer/cqn').Query} query
    * @returns {string} Database name
    */
   name(name) {
@@ -1148,11 +1145,6 @@ class CQN2SQLRenderer {
   managed_default(name, managed, src) {
     return `(CASE WHEN json_type(value,${this.managed_extract(name).extract.slice(8)}) IS NULL THEN ${managed} ELSE ${src} END)`
   }
-}
-
-// REVISIT: Workaround for JSON.stringify to work with buffers
-Buffer.prototype.toJSON = function () {
-  return this.toString('base64')
 }
 
 Readable.prototype[require('node:util').inspect.custom] = Readable.prototype.toJSON = function () { return this._raw || `[object ${this.constructor.name}]` }